chore: add release-please configuration file
This commit adds the release-please configuration file to the repository. The configuration file specifies the release type, package name, changelog path, and other settings for the release process. It also defines the sections for different types of changes in the changelog. This file will be used by the release-please tool to automate the release process.